# internal/handlers/docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact: {}
  title: Cryptocurrency Bot API documentation
  version: 0.1.0
paths:
  /account/new_account:
    get:
      consumes:
      - application/json
      description: Create account for new user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create account
      tags:
      - Account
...
  /account/portfolio/update:
    get:
      consumes:
      - application/json
      description: Update currency record in portfolio
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - Bearer: []
      summary: Update in portfolio
      tags:
      - Account - Portfolio
  /account/portfolio/view:
    get:
      consumes:
      - application/json
      description: View currencies in user portfolio
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - Bearer: []
      summary: View portfolio
      tags:
      - Account - Portfolio
...
  /account/price_observers/delete:
    get:
      consumes:
      - application/json
      description: Delete account price observers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete price observers
      tags:
      - Account - Price observers
...
  /info/index_rating:
    get:
      consumes:
      - application/json
      description: Get data to rating currency
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - Bearer: []
      summary: Currency rating
      tags:
      - Info - Rating
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"